requirements.txt
python-dotenv
agno
xpander-sdk[agno]
openai
anthropic
aioboto3
mcp
fastapi
uvicorn[standard]